requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
PyYAML==6.0.1
pytest==8.1.1